Refactoring / cleaning of the maintainer:
- Rename a number of internal components from `Migration` to
`Maintenance`, in order to reflect the fact that it is aimed at both
migrations and any other maintenance action that should occur in the
user Che account + OS namespace
- remove the shot version of arguments, which didn't have any
added-value and complicated the code
- Created some `utilities` functions, and refactored a bit the service
addount token manager to allow easier use in future endpoints.
- Add the ability to pass the OS master URL as a parameter (command-line
or in the REST endopint as query param or POST JSON value), instead of
using only an environment variable.
- Add a `debugRestEndpointsLocally.sh` file to start the server in debug
mode
- Add a `remote-debug.launch` debug configuration to easily start a
remote debug session on the Swarm application.